+++ autotest-watchr.rb
@@ -1,0 +1,9 @@
+
+def clear_console
+  puts "\e[H\e[2J"
+end
+
+watch(".*\.php") do |match|
+  clear_console
+  system 'phpunit'
+end