+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased][unreleased] - 2015-04-27
 ### Added
 - Added LICENSE file
+- Added watchr script for TDD
 
 ### Changed
 - Increased test coverage of Utils class